backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
motor==3.6.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
cryptography==43.0.1
python-multipart==0.0.9
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.1.1
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
httpx==0.27.2
orjson==3.10.7
python-dateutil==2.9.0.post0
